Finished renaming of solution.
