+++ README
@@ -1,4 +1,4 @@
-A circuit breaker implementation for .NET.
+A simple circuit breaker implementation for .NET.
 
 Forked and updated from Tatham Oddie's Reliability Patterns solution on GitHub:
 https://github.com/tathamoddie/reliability-patterns
