+++ SimpleCircuitBreaker.Tests/packages.config
