+++ SimpleCircuitBreaker/Properties/Resource.resx
