+++ SimpleCircuitBreaker/SimpleCircuitBreaker.nuspec
